backend/requirements.txt
cffi==2.0.0
cryptography==46.0.3
greenlet==3.2.4
mysql-connector-python==9.5.0
pillow==12.0.0
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
